requirements.txt
pandas
numpy
matplotlib
seaborn
statsmodels
pymc3
arviz
scikit-learn
